+++ examples/feature_showcase/source/gltf_model_test/gltf_model_test.cpp
@@ -104,7 +104,7 @@ std::unique_ptr<Error> GLTFModelTest::initialize()
 
         std::tie(vertexBuffer1, err) = graphicsDevice->createVertexBuffer(
             verticesCombo.data(),
-            verticesCombo.size(),
+            static_cast<u32>(verticesCombo.size()),
             sizeof(VertexCombined),
             gpu::BufferUsage::Immutable);
 
@@ -143,7 +143,7 @@ std::unique_ptr<Error> GLTFModelTest::initialize()
 
         std::tie(vertexBuffer2, err) = graphicsDevice->createVertexBuffer(
             verticesCombo.data(),
-            verticesCombo.size(),
+            static_cast<u32>(verticesCombo.size()),
             sizeof(VertexCombined),
             gpu::BufferUsage::Immutable);
 
@@ -171,7 +171,7 @@ std::unique_ptr<Error> GLTFModelTest::initialize()
         std::tie(indexBuffer, err) = graphicsDevice->createIndexBuffer(
             gpu::IndexFormat::UInt16,
             indices.data(),
-            indices.size(),
+            static_cast<u32>(indices.size()),
             gpu::BufferUsage::Immutable);
 
         if (err != nullptr) {
